Simplify vkrun

+++ sim/vkrun.py
@@ -29,12 +29,10 @@ import time
 from vkrun import parse
 from vkrun.entry import SeedGenerator, get_base_name, ExpandedEntry
 
-ansi_formatting = platform.system() in ["Darwin", "Linux"] and sys.stdout.isatty()
-log_stream = None
-run_exit = False
+run_exit = False   # signal to end run
+log_stream = None  # stream to log standard output
 
 
-# noinspection PyUnusedLocal
 def shutdown_handler(signum, frame):
     global run_exit
     run_exit = True
@@ -44,14 +42,18 @@ signal.signal(signal.SIGTERM, shutdown_handler)
 signal.signal(signal.SIGINT, shutdown_handler)
 
 
+build_dir = os.path.join(os.getcwd(), "build")
+vkbuild_dir = ""
+vkrun_dir = os.path.join(build_dir, "vkrun", datetime.now().strftime("%y%m%d_%H%M%S"))
+
+
 def main():
-    global log_stream
+    time_start = time.time()
+    if not os.path.exists(build_dir):
+        raise FileNotFoundError("build directory not found")
+
     parser = argparse.ArgumentParser()
-    parser.add_argument("-s", metavar="SIM", help="the simulator to target", default="xsim")
-    parser.add_argument("-b", metavar="BUILD", help="the input build directory", default="build/vkbuild")
     parser.add_argument("-t", metavar="TIMESTAMP", help="the build timestamp", default="")
-    parser.add_argument("-i", metavar="INPUT", help="the input rconf file", default="")
-    parser.add_argument("-o", metavar="OUTPUT", help="the output simulation directory", default="build/vkrun")
     parser.add_argument("-r", metavar="RANDSEED", help="the random number generator seed", type=int, default=0)
     parser.add_argument("-l", metavar="LOAD", help="the load factor", type=float, default=1)
     parser.add_argument("-k", metavar="KILL", help="the kill timeout", type=int, default=0)
@@ -59,93 +61,77 @@ def main():
     parser.add_argument("include", metavar="INCLUDE", help="entries to include", nargs="*", default=["none"])
     args = parser.parse_args()
 
-    time_start = time.time()
-    timestamp = datetime.now().strftime("%y%m%d_%H%M%S")
+    set_vkbuild_dir(args.t)
 
-    if args.s not in ["xsim"]:
-        raise ValueError("unsupported simulator %s" % args.s)
-
-    if args.t == "":
-        args.t = get_last_build(args.b)
-
-    build_dir = os.path.abspath(os.path.join(args.b, args.t))
-    output_dir = os.path.abspath(os.path.join(args.o, timestamp))
-
-    # log standard output
     if not args.d:
-        os.makedirs(output_dir, exist_ok=True)
-        log_file = os.path.join(output_dir, "vkrun.log")
-        log_stream = open(log_file, "w")
+        os.makedirs(vkrun_dir, exist_ok=True)
+        global log_stream
+        log_stream = open(os.path.join(vkrun_dir, "vkrun.log"), "w")
 
     if "none" in args.include:
         if len(args.include) != 1:
             raise ValueError("none cannot be used with other entries")
-        launch_none(args, timestamp, build_dir, output_dir)
+        launch_none(args)
     else:
-        launch_rconf(args, timestamp, build_dir, output_dir)
+        launch_rconf(args)
 
     time_end = time.time()
     elapsed = math.ceil(time_end - time_start)
     print_log()
     print_log("run complete in %s" % get_elapsed_string(elapsed))
     print_log()
-    if not args.d:
-        log_stream.close()
 
 
-def get_last_build(build_dir):
-    dirs = os.listdir(build_dir)
-    passing_dirs = []
-    for name in dirs:
-        path = os.path.join(build_dir, name)
-        if os.path.isdir(path) and os.path.exists(os.path.join(path, "PASS")):
-            passing_dirs.append(name)
-    passing_dirs = sorted(passing_dirs)
-    if not passing_dirs:
-        raise ValueError("no passing build directories found")
-    return passing_dirs[-1]
+def set_vkbuild_dir(timestamp):
+    global vkbuild_dir
+    if timestamp != "":
+        vkbuild_dir = os.path.join(build_dir, "vkbuild", timestamp)
+        if not os.path.exists(vkbuild_dir):
+            raise FileNotFoundError("build directory with timestamp %s not found" % timestamp)
+    else:
+        passing_dirs = []
+        for timestamp in os.listdir(os.path.join(build_dir, "vkbuild")):
+            path = os.path.join(build_dir, "vkbuild", timestamp)
+            if os.path.isdir(path) and os.path.exists(os.path.join(path, "PASS")):
+                passing_dirs.append(path)
+        passing_dirs = sorted(passing_dirs)
+        if not passing_dirs:
+            raise ValueError("no passing build directories found")
+        vkbuild_dir = passing_dirs[-1]
 
 
-def launch_none(args, timestamp, build_dir, output_dir):
+def launch_none(args):
     print_log()
     print_log("VKRUN")
     print_log("usage:  vkrun %s" % (" ".join(sys.argv[1:])))
-    print_log("build:  %s/%s" % (os.path.abspath(args.b), args.t))
-    print_log("run:    %s/%s" % (os.path.abspath(args.o), timestamp))
-    print_log("sim:    %s" % args.s)
+    print_log("build:  %s" % os.path.relpath(vkbuild_dir, build_dir))
+    print_log("run:    %s" % os.path.relpath(vkrun_dir, build_dir))
     print_log("kill:   %s" % args.k)
-    print_log("base:   none")
     print_log("count:  1")
     print_log()
 
     if args.d:
-        print_log("%s none" % get_label(0, 1))
+        log_result(0, 1, "none", True, 0)
     else:
-        run(build_dir, output_dir, "none", args.s, args.k, 1, ExpandedEntry("none", None, None, None), 0)
+        run(ExpandedEntry("none", None, None, None), "none", 0, 1, args.k)
 
 
-def launch_rconf(args, timestamp, build_dir, output_dir):
-    executor_workers = os.cpu_count()
-    if args.s == "xsim":
-        executor_workers = 1
-
+def launch_rconf(args):
     if args.l < 0:
         raise ValueError("load factor must be larger than 0")
     elif args.l > 100:
         raise ValueError("load factor must be smaller than 100")
 
-    if args.i == "":
-        args.i = os.path.join(args.b, args.t, "verik/rconf.txt")
-    rconf_file = os.path.abspath(args.i)
-
-    # copy rconf file
+    rconf_file = os.path.join(vkbuild_dir, "verik", "rconf.txt")
+    if not os.path.exists(rconf_file):
+        raise FileNotFoundError("runtime configuration file not found")
     if not args.d:
-        shutil.copyfile(rconf_file, os.path.join(output_dir, "rconf.txt"))
+        shutil.copyfile(rconf_file, os.path.join(vkrun_dir, "rconf.txt"))
 
     # parse rconf entries
     entries = parse.parse(rconf_file)
-    for name in args.include:
-        parse.include(entries, name)
+    for entry in args.include:
+        parse.include(entries, entry)
     seed_gen = SeedGenerator(args.r)
     expanded_entries = []
     for entry in entries:
@@ -155,10 +141,8 @@ def launch_rconf(args, timestamp, build_dir, output_dir):
     print_log()
     print_log("VKRUN")
     print_log("usage:  vkrun %s" % (" ".join(sys.argv[1:])))
-    print_log("build:  %s/%s" % (os.path.abspath(args.b), args.t))
-    print_log("rconf:  %s" % os.path.abspath(args.i))
-    print_log("run:    %s/%s" % (os.path.abspath(args.o), timestamp))
-    print_log("sim:    %s" % args.s)
+    print_log("build:  %s" % os.path.relpath(vkbuild_dir, build_dir))
+    print_log("run:    %s" % os.path.relpath(vkrun_dir, build_dir))
     print_log("seed:   %s" % args.r)
     print_log("load:   %d%%" % (int(args.l * 100)))
     print_log("kill:   %s" % args.k)
@@ -168,107 +152,104 @@ def launch_rconf(args, timestamp, build_dir, output_dir):
 
     for count, expanded_entry in enumerate(expanded_entries):
         if args.d:
-            print_log("%s %s" % (get_label(count, len(expanded_entries)), expanded_entry.name))
+            log_result(count, len(expanded_entries), expanded_entry.name, True, 0)
         else:
-            run(build_dir, output_dir, base_name, args.s, args.k, len(expanded_entries), expanded_entry, count)
+            run(expanded_entry, base_name, count, len(expanded_entries), args.k)
 
 
-def run(build_dir, output_dir, base_name, sim, timeout, total_count, expanded_entry, count):
-    global run_exit
+def run(expanded_entry, base_name, count, total_count, kill):
     time_start = time.time()
-    result = True
     relative_name = expanded_entry.relative_name(base_name)
-    sim_dir = os.path.join(output_dir, relative_name)
-    os.makedirs(sim_dir, exist_ok=True)
-    os.chdir(sim_dir)
+    entry_dir = os.path.join(vkrun_dir, relative_name)
+    os.makedirs(entry_dir, exist_ok=True)
+    os.chdir(entry_dir)
     if run_exit:
         result = False
     else:
         try:
-            if sim == "xsim":
-                run_xsim(build_dir, sim_dir, timeout)
+            run_xsim(entry_dir, kill)
         except:
             result = False
+        else:
+            result = True
     if result:
-        open(os.path.join(sim_dir, "PASS"), "w").close()
+        open(os.path.join(entry_dir, "PASS"), "w").close()
     else:
-        open(os.path.join(sim_dir, "FAIL"), "w").close()
+        open(os.path.join(entry_dir, "FAIL"), "w").close()
     time_end = time.time()
     elapsed = int(math.ceil(time_end - time_start))
-    log_result(total_count, count, expanded_entry.name, result, elapsed)
+    log_result(count, total_count, expanded_entry.name, result, elapsed)
 
 
-def run_xsim(input_dir, sim_dir, timeout):
+def run_xsim(entry_dir, kill):
     time_start = time.time()
-    ln_target = os.path.join(input_dir, "xsim.dir", "sim")
-    ln_dir = os.path.join(sim_dir, "xsim.dir")
-    os.makedirs(ln_dir, exist_ok=True)
+    os.makedirs(os.path.join(entry_dir, "xsim.dir"), exist_ok=True)
+    xsim_build = os.path.join(vkbuild_dir, "xsim.dir", "sim")
+    xsim_entry = os.path.join(entry_dir, "xsim.dir", "sim")
     devnull = open(os.devnull, "w")
     if platform.system() == "Windows":
-        subprocess.run(["mklink", "/D", os.path.join(ln_dir, "sim"), ln_target], shell=True, check=True, stdout=devnull)
+        subprocess.run(["mklink", "/D", xsim_entry, xsim_build], shell=True, check=True, stdout=devnull)
     else:
-        subprocess.run(["ln", "-s", ln_target, ln_dir], check=True)
+        subprocess.run(["ln", "-s", xsim_build, xsim_entry], check=True)
     process = subprocess.Popen([shutil.which("xsim"), "-R", "sim"], stdout=devnull)
     while process.poll() is None:
         time.sleep(1)
         time_current = time.time()
-        if 0 < timeout < time_current - time_start:
+        if 0 < kill < time_current - time_start:
             process.kill()
-            raise RuntimeError("timeout of %d exceeded" % timeout)
+            raise RuntimeError("kill timeout of %d exceeded" % kill)
         if run_exit:
             process.kill()
             raise RuntimeError("exit signal asserted")
 
 
-def log_result(total_count, count, name, result, elapsed):
+def log_result(count, total_count, name, result, elapsed):
+    label = get_label(count, total_count, result, elapsed)
+    if name != "none":
+        name_string = " %s" % name
+    else:
+        name_string = ""
     if result:
         color_string = u"\u001B[32m\u001B[1m"  # ANSI green bold
-        color_string += get_label(count, total_count, result, elapsed)
+        color_string += label
         color_string += u"\u001B[0m"  # ANSI reset
         color_string += u"\u001B[32m"  # ANSI green
-        color_string += " %s" % name
+        color_string += name_string
         color_string += u"\u001B[0m"  # ANSI reset
     else:
         color_string = u"\u001B[31m\u001B[1m"  # ANSI red bold
-        color_string += get_label(count, total_count, result, elapsed)
+        color_string += label
         color_string += u"\u001B[0m"  # ANSI reset
         color_string += u"\u001B[31m"  # ANSI red
-        color_string += " %s" % name
+        color_string += name_string
         color_string += u"\u001B[0m"  # ANSI reset
-    plain_string = get_label(count, total_count, result, elapsed)
-    plain_string += " %s" % name
+    plain_string = label + name_string
     print_log(color_string, plain_string)
 
 
-def get_label(count, total_count, result=None, elapsed=None):
+def get_label(count, total_count, result, elapsed):
     count_string = str(count + 1).zfill(len(str(total_count)))
-    if result is None:
-        return "[%s/%d]" % (count_string, total_count)
-    elif result:
+    if result:
         return "[%s/%d PASS %s]" % (count_string, total_count, get_elapsed_string(elapsed))
     else:
         return "[%s/%d FAIL %s]" % (count_string, total_count, get_elapsed_string(elapsed))
 
 
 def get_elapsed_string(elapsed):
-    if elapsed is None:
-        return "00:00:00"
-    else:
-        seconds = elapsed % 60
-        elapsed /= 60
-        minutes = elapsed % 60
-        elapsed /= 60
-        hours = elapsed
-        return "%02d:%02d:%02d" % (hours, minutes, seconds)
+    seconds = elapsed % 60
+    elapsed /= 60
+    minutes = elapsed % 60
+    elapsed /= 60
+    hours = elapsed
+    return "%02d:%02d:%02d" % (hours, minutes, seconds)
 
 
 def print_log(color_string=None, plain_string=None):
-    global log_stream
     if color_string is None:
         color_string = ""
     if plain_string is None:
         plain_string = color_string
-    if ansi_formatting:
+    if platform.system() in ["Darwin", "Linux"] and sys.stdout.isatty():
         print(color_string)
     else:
         print(plain_string)
